Flyttet MeldFraEndringer til Felles.kt og rettet feil i NN teksten
